disable tailwind prose by default
- abstracted prose to a separate file

This is because I can't figure out how to only get the prose content and
only purge it in certain envs. These styles are only used in the blog
component so there's no reason to include them in the marketing component

Hopefully in the future this can be done, because extracting the prose
styles from the build is not trivial.